Add content to HistogramNode in entwurfsheft. Also changed the type of Data in the HistogramNode class to double, to reflect the relative distribution of color values

+++ implementation/YuvKA/Pipeline/Implementation/HistogramNode.cs
@@ -22,7 +22,7 @@ namespace YuvKA.Pipeline.Implementation
 		}
 
 		[Browsable(false)]
-		public int[] Data { get; private set; }
+		public double[] Data { get; private set; }
 
 		public override Frame[] ProcessFrame(Frame[] inputs, int frameIndex)
 		{
